requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.38.18
botocore==1.38.18
celery==5.5.2
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.6
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
git-filter-repo==2.47.0
idna==3.10
jmespath==1.0.1
kombu==5.5.3
msgpack==1.1.0
multidict==6.4.4
olefile==0.47
phonenumbers==9.0.5
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==6.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
twilio==9.6.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
gunicorn==21.2.0
